# backend/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Artifex Gallery API",
    "version": "1.0.0",
    "description": "AI-powered image and video gallery with auto-tagging, NSFW detection, and federation support."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string", "minLength": 3 }, "password": { "type": "string", "minLength": 8 }, "display_name": { "type": "string" } } } } } },
        "responses": { "201": { "description": "User created" }, "400": { "description": "Validation error" }, "409": { "description": "Username taken" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get JWT token",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Login successful, returns token" }, "401": { "description": "Invalid credentials" }, "403": { "description": "Account disabled" } }
      }
    },
...
    "/images/upload": {
      "post": {
        "tags": ["Images"], "summary": "Upload images/videos (up to 50)", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "skip_duplicates", "in": "query", "schema": { "type": "string", "default": "true" } }],
        "requestBody": { "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "images": { "type": "array", "items": { "type": "string", "format": "binary" } } } } } } },
        "responses": { "201": { "description": "Upload results with duplicate detection" }, "413": { "description": "Storage quota exceeded" } }
      }
    },
    "/images": {
      "get": {
        "tags": ["Images"], "summary": "List all visible images (public + own private)",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "model", "in": "query", "schema": { "type": "string" } },
          { "name": "sampler", "in": "query", "schema": { "type": "string" } },
          { "name": "media_type", "in": "query", "schema": { "type": "string", "enum": ["image", "video"] } },
          { "name": "tag", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest", "oldest"] } }
        ],
        "responses": { "200": { "description": "Paginated image list" } }
      }
    },
...
    "/collections": {
      "get": { "tags": ["Collections"], "summary": "List collections", "responses": { "200": { "description": "Collection list" } } },
      "post": {
        "tags": ["Collections"], "summary": "Create collection", "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/tags": { "get": { "tags": ["Tags"], "summary": "List all tags with counts", "parameters": [{ "name": "category", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Tag list" } } } },
...
  }
}